backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
django-heroku==0.3.1
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.9
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-writable-nested==0.6.3
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
psycopg2==2.9.3
pygraphviz==1.9
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
zipp==3.7.0
